Add check image

+++ app/article/routes.py
@@ -32,7 +32,8 @@ def add():
 @login_required
 def delete(id):
 	article = Article.query.filter_by(id=id).first_or_404()
-	os.remove(os.path.join(current_app.root_path, 'static/images', article.image))
+	if article.image is not None;:
+		os.remove(os.path.join(current_app.root_path, 'static/images', article.image))
 	db.session.delete(article)
 	db.session.commit()
 	flash('Your article has been removed!', 'success')
